[ci bootstrap] Do not add external gateway to router
For clouds where instances directly attached to public
network, having a router with default gateway is unnecessary
as secondary network is mostly used for internal traffic.
This will save some external IPs i.e 1 per job.

This needs to be set only when the test vm needs to have
external connectivity which is not needed in most cases.

Depends-On: https://review.rdoproject.org/r/c/config/+/57579
Related-Issue: OSPCIX-771