Add OpenStack Operator Initialization
We have introduced new tasks for initializing the OpenStack
operator. This is required for versions prior to v1.0.7, where
deployment didn't include the initialization. This action only takes
place if we are using an OLM-based update.

Additionally, we block the process until the OpenStack deployment is
successful. This applies to all starting versions, as the operator
should be ready before proceeding.

Closes: [OSPRH-15799](https://issues.redhat.com/browse/OSPRH-15799)